auto-internet-speed-checker/requirements.txt
speedtest-cli==2.1.3
pandas==2.2.2
matplotlib==3.9.0